+++ Cryptex.php
@@ -1,29 +1,32 @@
 <?php
 /**
     Plugin Name: Cryptex - E-Mail Address Protection
-    Plugin URI: https://andidittrich.de/go/cryptex
+    Plugin URI: https://github.com/AndiDittrich/WordPress.Cryptex
     Description: Advanced Graphical E-Mail Obfuscator which provides image based email address protection using wordpress shortcode and integrated encryption/decryption of addresses
-    Version: 6.1-BETA1
+    Version: 7.0-BETA1
     Author: Andi Dittrich
     Author URI: https://andidittrich.de
     License: MIT X11 License
+    Text Domain: cryptex
+    Domain Path: /lang
 
     ----
     The MIT License (X11 License)
-    Copyright (c) 2010-2017 Andi Dittrich <https://andidittrich.de>
+    Copyright (c) 2010-2018 Andi Dittrich <https://andidittrich.de>
     Permission is hereby granted, free of charge, to any personobtaining a copy of this software and associated documentationfiles (the "Software"), to deal in the Software withoutrestriction, including without limitation the rights to use,copy, modify, merge, publish, distribute, sublicense, and/or sellcopies of the Software, and to permit persons to whom theSoftware is furnished to do so, subject to the followingconditions:
     The above copyright notice and this permission notice shall beincluded in all copies or substantial portions of the Software.
     THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIESOF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE ANDNONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHTHOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISINGFROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OROTHER DEALINGS IN THE SOFTWARE.
 */
 
 
-/*  AUTO GENERATED FILE - DO NOT EDIT !!
-    WP-SKELEKTON | MIT X11 License | https://github.com/AndiDittrich/WP-Skeleton
-    ------------------------------------
-    Plugin Bootstrap Operation
-*/
+// Plugin Bootstrap Operation
+// AUTO GENERATED CODE - DO NOT EDIT !!!
+
 define('CRYPTEX_INIT', true);
-define('CRYPTEX_VERSION', '6.1-BETA1');
+define('CRYPTEX_VERSION', '7.0-BETA1');
+define('CRYPTEX_WPSKLTN_VERSION', '0.13.0');
+define('CRYPTEX_PLUGIN_TITLE', 'Cryptex - E-Mail Address Protection');
+define('CRYPTEX_PLUGIN_HEADLINE', 'Advanced Graphical E-Mail Obfuscator which provides image based email address protection using wordpress shortcode and integrated encryption/decryption of addresses');
 define('CRYPTEX_PLUGIN_PATH', dirname(__FILE__));
 define('CRYPTEX_PLUGIN_URL', plugins_url('/cryptex/'));
 define('CRYPTEX_DEFAULT_FONT_PATH', CRYPTEX_PLUGIN_PATH.DIRECTORY_SEPARATOR.'fonts'.DIRECTORY_SEPARATOR);