+++ class/ResourceLoader.php
@@ -17,74 +17,62 @@
 */
 namespace Cryptex;
 
+use \Cryptex\skltn\ResourceManager as ResourceManager;
+
 class ResourceLoader{
         
     // stores the plugin config
-    private $_config;
+    private $_settingsManager;
     
     private $_cacheManager;
 
-    public function __construct($settingsUtil, $cacheManager){
+    public function __construct($settingsManager, $cacheManager){
         // store local plugin config
-        $this->_config = $settingsUtil->getOptions();
+        $this->_settingsManager = $settingsManager;
         $this->_cacheManager = $cacheManager;
     }
     
-    /**
-     * Apply frontend hooks
-     */
+    // frontend resources
     public function frontend(){
         // load frontend css
-        if ($this->_config['embed-css']){
-            add_action('wp_head', array($this, 'appendInlineCSS'), 50);
+        if ($this->_settingsManager->getOption('embed-css')){
+            // get css
+            $css = file_get_contents($this->_cacheManager->getCachePath().'Cryptex.css');
+
+            // add content to enqeue cache
+            ResourceManager::enqueueDynamicStyle($css);
         }
         
         // only include js if required
-        if ($this->_config['enable-hyperlink'] && $this->_config['embed-js']){
-            add_action('wp_footer', array($this, 'appendInlineJavascript'));
+        if ($this->_settingsManager->getOption('enable-hyperlink') && $this->_settingsManager->getOption('embed-js')){
+            // get js
+            $js = file_get_contents(CRYPTEX_PLUGIN_PATH.'/resources/CryptexHyperlinkDecoder.min.js');
+
+            // drop trailing function call "();"
+            $js = substr($js, 0, -3);
+
+            // append initialization - function call with arguments
+            $js .= '(window, document, "' . KeyShiftingEncoder::getKey() . '");';
+            
+            // add content to enqeue cache
+            ResourceManager::enqueueDynamicScript($js);
         }
-    }
-
-    // inline js
-    public function appendInlineJavascript(){
-        // get js
-        $js = file_get_contents(CRYPTEX_PLUGIN_PATH.'/resources/CryptexHyperlinkDecoder.min.js');
-
-        // drop trailing ();
-        $js = substr($js, 0, -3);
-
-        // output inline scripts - add function call
-        echo '<script type="text/javascript">/* <![CDATA[ */', $js, '(window, document, "', KeyShiftingEncoder::getKey() , '"); /* ]]> */</script>';
-    }
-
-    // inline css
-    public function appendInlineCSS(){
-        echo '<style type="text/css">', file_get_contents($this->_cacheManager->getCachePath().'Cryptex.css'), '</style>';
     }
     
     public function appendAdminCSS(){
         // colorpicker css
-        wp_register_style('cryptex-jquery-colorpicker', plugins_url('/cryptex/resources/extern/colorpicker/css/colorpicker.css'));
-        wp_enqueue_style('cryptex-jquery-colorpicker');
+        ResourceManager::enqueueStyle('cryptex-jquery-colorpicker', 'extern/colorpicker/css/colorpicker.css');
         
         // new UI !
-        wp_register_style('cryptex-settings', plugins_url('/cryptex/resources/admin/CryptexSettings.css'));
-                
-        // settings css        
-        wp_enqueue_style('cryptex-settings');
+        ResourceManager::enqueueStyle('cryptex-settings', 'admin/settings.css');
     }
     
     public function appendAdminJS(){
-        // jquery cookie js
-        wp_register_script('cryptex-jquery-cookie', plugins_url('/cryptex/resources/extern/jquery.cookie/jquery.cookie.js'), array('jquery'));
-        wp_enqueue_script('cryptex-jquery-cookie');
 
         // colorpicker js
-        wp_register_script('cryptex-jquery-colorpicker', plugins_url('/cryptex/resources/extern/colorpicker/js/colorpicker.js'), array('jquery'));
-        wp_enqueue_script('cryptex-jquery-colorpicker');
-        
+        ResourceManager::enqueueScript('cryptex-jquery-colorpicker', 'extern/colorpicker/js/colorpicker.js', array('jquery'));
+
         // settings init script
-        wp_register_script('cryptex-settings-init', plugins_url('/cryptex/resources/admin/CryptexSettings.min.js'), array('jquery'));
-        wp_enqueue_script('cryptex-settings-init');
+        ResourceManager::enqueueScript('cryptex-settings', 'admin/settings.js', array('jquery', 'jquery-color'));
     }
 }