+++ skltn/CssBuilder.php
@@ -1,9 +1,19 @@
 <?php
-/*  AUTO GENERATED FILE - DO NOT EDIT !!
-    WP-SKELEKTON | MIT X11 License | https://github.com/AndiDittrich/WP-Skeleton
-    ------------------------------------
-    Generate Dynamic CSS Files
-*/
+// ---------------------------------------------------------------------------------------------------------------
+// -- WP-SKELETON AUTO GENERATED FILE - DO NOT EDIT !!!
+// --
+// -- Copyright (c) 2016-2018 Andi Dittrich
+// -- https://github.com/AndiDittrich/WP-Skeleton
+// --
+// ---------------------------------------------------------------------------------------------------------------
+// --
+// -- This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+// -- If a copy of the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+// --
+// ---------------------------------------------------------------------------------------------------------------
+
+// Generate Dynamic CSS Files
+
 namespace Cryptex\skltn;
 
 class CssBuilder{
@@ -30,7 +40,7 @@ class CssBuilder{
     // render ruleset
     public function render(){
         // local output buffer
-        $css = '/* Cryptex Dynamical Generated Stylesheet - DO NOT EDIT */';
+        $css = '/* Cryptex dynamic generated stylesheet - DO NOT EDIT */';
 
         // get rulesets
         foreach ($this->_buffer as $selector => $rules){